fixed gpio_write and gpio_read, issue with redefining serial_init

+++ include/gpio.h
@@ -29,7 +29,7 @@ void _gpio_init(gpio_port port, uint32_t* pins, size_t pinnum, bool output, uint
 // hacky workaround to not require user to specifiy pinnum
 #define gpio_init(port, pin_array, output, alt_func_select, analogue_select, interruptToggle) _gpio_init(port, pin_array, len(pin_array), output, alt_func_select, analogue_select, interruptToggle)
 
-uint8_t gpio_read(gpio_port, int32_t pin, ...);
-uint8_t gpio_write(gpio_port, uint8_t data);
+uint32_t gpio_read(gpio_port, int32_t pin, ...);
+uint32_t gpio_write(gpio_port, uint32_t data);
 
 #endif /* INCLUDE_GPIO_H_ */
